Added a DeveloperSettings class which is "Editor Only" but based on a project...
This is for testing adding CSV files that can be used later for data (in this test case...String Tables). But with a DetailsCustomization since CallInEditor UPROPERTY doesn't work in DeveloperSettings...

The DetailsCustomization (introduced here ) is also a way to "validate" the filename like what the Content Browser does!

SUPER WIP :P